+++ src/components/form/DefaultFormContents.tsx
@@ -1,50 +1,37 @@
 import i18n from '@dhis2/d2-i18n'
-import { NoticeBox } from '@dhis2/ui'
-import React, { useEffect, useRef } from 'react'
+import { Card } from '@dhis2/ui'
+import React from 'react'
+import { useFormState } from 'react-final-form'
 import { useNavigate } from 'react-router-dom'
 import { getSectionPath } from '../../lib'
 import { ModelSection } from '../../types'
-import { StandardFormSection, StandardFormActions } from '../standardForm'
+import { StandardFormActions, StandardFormSection } from '../standardForm'
 import classes from './DefaultFormContents.module.css'
+import { DefaultFormErrorNotice } from './DefaultFormErrorNotice'
 
-export function DefaultFormContents({
+export function DefaultEditFormContents({
     children,
     section,
-    submitError,
-    submitting,
 }: {
     children: React.ReactNode
     section: ModelSection
-    submitting: boolean
-    submitError?: string
 }) {
-    const formErrorRef = useRef<HTMLDivElement | null>(null)
+    const { submitting } = useFormState({
+        subscription: { submitting: true },
+    })
     const navigate = useNavigate()
 
     const listPath = `/${getSectionPath(section)}`
-    useEffect(() => {
-        if (submitError) {
-            formErrorRef.current?.scrollIntoView({ behavior: 'smooth' })
-        }
-    }, [submitError])
 
     return (
         <>
-            <div className={classes.form}>{children}</div>
-            {submitError && (
+            <Card className={classes.form}>
+                {children}
+
                 <StandardFormSection>
-                    <div ref={formErrorRef}>
-                        <NoticeBox
-                            error
-                            title={i18n.t(
-                                'Something went wrong when submitting the form'
-                            )}
-                        >
-                            {submitError}
-                        </NoticeBox>
-                    </div>
+                    <DefaultFormErrorNotice />
                 </StandardFormSection>
-            )}
+            </Card>
             <div className={classes.formActions}>
                 <StandardFormActions
                     cancelLabel={i18n.t('Cancel')}
@@ -54,5 +41,38 @@ export function DefaultFormContents({
                 />
             </div>
         </>
+    )
+}
+
+export function DefaultNewFormContents({
+    section,
+    children,
+}: {
+    children: React.ReactNode
+    section: ModelSection
+}) {
+    const { submitting } = useFormState({
+        subscription: { submitting: true },
+    })
+
+    const navigate = useNavigate()
+
+    const listPath = `/${getSectionPath(section)}`
+
+    return (
+        <Card className={classes.form}>
+            {children}
+            <StandardFormSection>
+                <DefaultFormErrorNotice />
+            </StandardFormSection>
+            <StandardFormActions
+                cancelLabel={i18n.t('Exit without saving')}
+                submitLabel={i18n.t('Create {{modelName}} ', {
+                    modelName: section.title,
+                })}
+                submitting={submitting}
+                onCancelClick={() => navigate(listPath)}
+            />
+        </Card>
     )
 }