+++ src/components/sectionList/filters/filterSelectors/ConstantFilters.tsx
@@ -1,11 +1,17 @@
 import React from 'react'
-import { AGGREGATION_TYPE, DOMAIN_TYPE, VALUE_TYPE } from '../../../../lib'
+import {
+    AGGREGATION_TYPE,
+    DATA_DIMENSION_TYPE,
+    DOMAIN_TYPE,
+    getTranslatedProperty,
+    VALUE_TYPE,
+} from '../../../../lib'
 import { ConstantSelectionFilter } from './ConstantSelectionFilter'
 
 export const DomainTypeSelectionFilter = () => {
     return (
         <ConstantSelectionFilter
-            label={'Domain type'}
+            label={getTranslatedProperty('domainType')}
             filterKey="domainType"
             constants={DOMAIN_TYPE}
         />
@@ -15,7 +21,7 @@ export const DomainTypeSelectionFilter = () => {
 export const ValueTypeSelectionFilter = () => {
     return (
         <ConstantSelectionFilter
-            label={'Value type'}
+            label={getTranslatedProperty('valueType')}
             filterKey="valueType"
             constants={VALUE_TYPE}
             filterable
@@ -26,10 +32,20 @@ export const ValueTypeSelectionFilter = () => {
 export const AggregationTypeFilter = () => {
     return (
         <ConstantSelectionFilter
-            label={'Aggregation type'}
+            label={getTranslatedProperty('aggregationType')}
             filterKey="aggregationType"
             constants={AGGREGATION_TYPE}
             filterable
+        />
+    )
+}
+
+export const DataDimensionTypeFilter = () => {
+    return (
+        <ConstantSelectionFilter
+            label={getTranslatedProperty('dataDimensionType')}
+            filterKey="dataDimensionType"
+            constants={DATA_DIMENSION_TYPE}
         />
     )
 }