+++ src/lib/sectionList/filters/filterConfig.tsx
@@ -1,6 +1,6 @@
 import { StringParam } from 'use-query-params'
 import { z } from 'zod'
-import { DataElement } from '../../../types/generated'
+import { Category, DataElement } from '../../../types/generated'
 import { IDENTIFIABLE_FILTER_KEY } from '../../constants'
 import { isValidUid, parseAccessString } from '../../models'
 import { CustomDelimitedArrayParam } from './customParams'
@@ -13,12 +13,14 @@ export const filterParamsSchema = z
         [IDENTIFIABLE_FILTER_KEY]: z.string(),
         aggregationType: z.array(z.nativeEnum(DataElement.aggregationType)),
         categoryCombo: zodArrayIds,
+        category: zodArrayIds,
         dataSet: zodArrayIds,
         domainType: z.array(z.nativeEnum(DataElement.domainType)),
         publicAccess: z.array(
             z.string().refine((val) => parseAccessString(val) !== null)
         ),
         valueType: z.array(z.nativeEnum(DataElement.valueType)),
+        dataDimensionType: z.nativeEnum(Category.dataDimensionType),
     })
     .partial()
 
@@ -30,8 +32,10 @@ export const filterQueryParamType = {
     domainType: CustomDelimitedArrayParam,
     valueType: CustomDelimitedArrayParam,
     dataSet: CustomDelimitedArrayParam,
+    category: CustomDelimitedArrayParam,
     categoryCombo: CustomDelimitedArrayParam,
     publicAccess: CustomDelimitedArrayParam,
+    dataDimensionType: StringParam,
 } as const satisfies QueryParamsConfigMap
 
 export const validFilterKeys = Object.keys(filterQueryParamType)
@@ -57,4 +61,5 @@ export type ConfigurableFilterKey = Exclude<
     FilterKey,
     typeof IDENTIFIABLE_FILTER_KEY
 >
+
 export type FilterKeys = FilterKey[]