+++ src/lib/zod/index.ts
@@ -1,1 +1,2 @@
 export { segmentsToPath } from './segmentsToPath'
+export { getDefaults } from './getDefaults'