+++ src/pages/dataElementGroupSets/Edit.tsx
@@ -6,9 +6,10 @@ import { withTypes } from 'react-final-form'
 import { useNavigate, useParams } from 'react-router-dom'
 import { Loader } from '../../components'
 import {
-    DefaultFormContents,
+    DefaultEditFormContents,
     useCustomAttributesQuery,
 } from '../../components/form'
+import { AttributeMetadata } from '../../components/form/attributes/useCustomAttributesQuery'
 import {
     SCHEMA_SECTIONS,
     getSectionPath,
@@ -17,7 +18,7 @@ import {
 } from '../../lib'
 import { createJsonPatchOperations } from '../../lib/form/createJsonPatchOperations'
 import { getAllAttributeValues } from '../../lib/models/attributes'
-import { Attribute, DataElementGroupSet } from '../../types/generated'
+import { DataElementGroupSet } from '../../types/generated'
 import {
     DataElementGroupSetFormFields,
     dataElementGroupSetSchema,
@@ -55,7 +56,7 @@ function computeInitialValues({
     customAttributes,
 }: {
     dataElementGroupSet: DataElementGroupSet
-    customAttributes: Attribute[]
+    customAttributes: AttributeMetadata[]
 }) {
     if (!dataElementGroupSet) {
         return {}
@@ -64,7 +65,7 @@ function computeInitialValues({
     // We want to have an array in the state with all attributes, not just the
     // ones that are set
     const attributeValues = getAllAttributeValues(
-        dataElementGroupSet,
+        dataElementGroupSet.attributeValues,
         customAttributes
     )
 
@@ -78,7 +79,7 @@ function computeInitialValues({
         dataDimension: dataElementGroupSet.dataDimension,
         dataElementGroups: dataElementGroupSet.dataElementGroups || [],
         attributeValues,
-    }
+    } as FormValues
 }
 
 export const Component = () => {
@@ -117,7 +118,7 @@ function DataElementGroupSetForm({
     attributes,
 }: {
     dataElementGroupSet: DataElementGroupSet
-    attributes: Attribute[]
+    attributes: AttributeMetadata[]
 }) {
     const navigate = useNavigate()
     const patchDirtyFields = usePatchModel(
@@ -152,15 +153,11 @@ function DataElementGroupSetForm({
                 customAttributes: attributes,
             })}
         >
-            {({ handleSubmit, submitting, submitError }) => (
+            {({ handleSubmit }) => (
                 <form onSubmit={handleSubmit}>
-                    <DefaultFormContents
-                        section={section}
-                        submitting={submitting}
-                        submitError={submitError}
-                    >
+                    <DefaultEditFormContents section={section}>
                         <DataElementGroupSetFormFields />
-                    </DefaultFormContents>
+                    </DefaultEditFormContents>
                 </form>
             )}
         </Form>